requirements.txt
APScheduler==3.9.1.post1
asyncpg==0.27.0
discord.py==2.1.0
psycopg2==2.9.5
SQLAlchemy==1.4.46